data_warehouse/ETL/requirements.txt
pandas==2.0.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
numpy==1.24.3